feat: task 8 -- extract solana module + httpmock tests + isolate verify_usdc_transfer
Move mcp/src/solana.rs to core/src/solana/mod.rs. Removed verify_usdc_transfer
from SolanaClient; it now lives in mcp/src/payment.rs as a standalone
`pub async fn verify_usdc_transfer(client: &SolanaClient, ...)`.

SolanaClient::rpc is now pub so the extracted fn can call it cross-crate.
Added 8 httpmock tests covering write_memo, read_memo (hit+miss), airdrop,
get_tx_signers, confirm_tx retry exhaustion (start_paused to avoid wall-clock),
health_check, and RPC error propagation. Added tokio test-util feature to
core/Cargo.toml dev-deps for paused-clock support.

Fixed a clippy::unnecessary_map_or on the moved confirm_tx (map_or → is_none_or).
Updated imports in mcp/src/{main.rs, mcp.rs, tools.rs, payment.rs} to use
mnemonic_core::solana::SolanaClient; the deposit handler in main.rs now calls
payment::verify_usdc_transfer(&state.solana, ...) instead of the former method.

Verification:
- cargo test -p mnemonic-core -- solana -> 8 passed
- grep -r "verify_usdc_transfer" core/src/ -> empty
- cargo clippy -p mnemonic-core -- -D warnings -> clean
- cargo build -p mnemonic-mcp -> success
