+++ work/mnemonic-core/tasks/8.md
@@ -1,5 +1,5 @@
 ---
-status: planned
+status: done
 depends_on: [7]
 wave: 8
 skills: [code-writing]
